pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-review-agent"
version = "1.0.0"
description = "A Python-based PR review agent for multiple git servers"
authors = [{name = "PR Review Agent", email = "pr-review@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "pylint>=2.17.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
]

[project.scripts]
pr-review = "pr_review_agent.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pylint]
max-line-length = 88
disable = ["C0114", "C0116"]
